requirements.txt

openai
langchain
chromadb
newspaper3k
streamlit
tiktoken
